feat: switch to dependecy injection
